Problem: x64 HDF5 libs would require alot more of StackSize and HeapSize, yielding StackOverflow whenever writing to hdf5 files. Solution: Substituted every minval and maxval array functions with standard do-loop functions created in ModuleFunctions (minival and maxival). Substituted every MatrixA = MatrixB type of code in ModuleHdf5 with do-loop function SetMatrixValue. This makes the code a lot more resistant to StackOverflow issues deriving from Stack and Heap use.
Problem2: In x64, during the KillMain, at the end of the run, an error would rise in the deallocate(Thomas%Coef3).
Solution2: Commented the deallocate of the type Coef3 (it's not a matrix but a type with pointers to matrices). It solved the problem. 

Now x64 runs without known problems.

Changed the STACKSIZE RESERVE to 640 000 000 and HEAPSIZE RESERVE to 1280000 in certain configurations of the x64 platform. This allows to run some models with Stack overflow issues.

git-svn-id: https://mohid.svn.codeplex.com/svn@62757 81fbe566-5dc4-48c1-bdea-7421811ca204