Found an insidious incoherence in Intel Fortran 11.1 in windows between openmp and non-openmp versions. Blogged about it here http://guillaume-maretec.blogspot.com/2011/04/always-nullify-new-pointers-before.html
This made me fix a bug in moduleMap that would only show up in the openmp version but that had nothing to do with openmp directives.

Basically, the rule of thumb for programming in fortran90 is to *always* nullify explicitly your new pointers after their declaration. 

Otherwise they could be considered to be *associated* when in fact they are *not*. For some reason, Intel Fortran 11.1, in windows, without openmp, is permissive and with openmp it is not.

Here's an example of what I'm  talking about:

type (T_MYTYPE), pointer     :: ImAPointer
nullify(ImAPointer) ! always nullify
...
!This test will be inconsistent between openmp and non-openmp 
!if the nullify directive above is omitted. BEWARE.
if (.not. associated(ImAPointer)
    !some code
else
    !some other code
endif


git-svn-id: https://mohid.svn.codeplex.com/svn@65160 81fbe566-5dc4-48c1-bdea-7421811ca204