Added a new fortran95 compiler in the linux makefile project, Open64 (http://www.open64.net/). It requires that the Fortran interface with the HDF5 library be compiled with openf95 as well.
The Open64 compiler required a few minor modifications to the Mohid project source files for it to build (not many, though).

Open64 is posix, unix, linux machines only (yet); but at least we can ensure that the MOHID coding pattern is more universal than intel fortran in order to meet requirements.

git-svn-id: https://mohid.svn.codeplex.com/svn@66736 81fbe566-5dc4-48c1-bdea-7421811ca204