Was missing hydrodynamic commit from last commit (codeplex server problem related...).
(Implemented and tested openmp directives in moduleOpenBoundary. Structural changes needed to be made to make calls to
other module subroutines, such as interpolation in moduleTriangulation, thread-safe.

The structural changes, alone, made the 2D hydrodynamic model, forced with several tide gauges that must be interpolated to the model boundary cells, run almost twice as fast in our father PCOMS model (of the portuguese coast).)

git-svn-id: https://mohid.svn.codeplex.com/svn@69220 81fbe566-5dc4-48c1-bdea-7421811ca204