Refactored a Parallelized Zone (PZ) in ModuleOpenBoundary. Created _ThreadSafe versions of public methods of the modules Triangulation, HorizontalGrid and Gauge, subroutines Ready_ThreadSafe, LocateObj_ThreadSafe, Get_xxx_ThreadSafe and Interpolation_ThreadSafe. Reverted the Ready subroutines in these modules to the standard ModuleShell version (cleaned some comments as well).
A _ThreadSafe version of a public method of a class module consists in not using the global auxiliary pointer variable Me, but rather using a local auxiliary pointer variable LocalMe. It also consists in declaring the outward intent of all its arguments (specifically intent(out) or intent(inout)). Finally, the state of the LocalMe data-structure cannot be changed. If these conditions are met, then the public method of the class module is considered to be _ThreadSafe and can be called safely in parallelized zones.

Also started documenting a section intitled parallel programming in the MOHID FrameWork Programmers Manual describing these issues ...

git-svn-id: https://mohid.svn.codeplex.com/svn@72188 81fbe566-5dc4-48c1-bdea-7421811ca204