Bug correction in subroutine OutPutTimeInternal (ModuleEnterData).
The position were the instruction "ReadValuesNumber = iflag" was inside the routine (after the verification of the number of instants provided by the user) was turnning off the correction of ReadValuesNumber to 2, when only one output time was provided (only regular outputs), forcing the use of at least 2 output times to avoid an exception.

The instruction was moved to before the verification.

git-svn-id: https://mohid.svn.codeplex.com/svn@72525 81fbe566-5dc4-48c1-bdea-7421811ca204