The two modules for seagrasses were included in the code.
Module SeagrassSedimInteraction calculates roots uptake of nutrients from the sediment.
After the uptake, the ammonia 
concentration is updated in the sediment column. 
This module is not calculating the sources-sinks terms
for roots. The  ModuleBenthicEcology calculates the sources-sinks terms
for roots.


Module SeagrassWaterInteraction 
calculates uptake of ammonia, nitrate and phosphate
in the water column by seagrasses leaves. It also calculates the 
light limitation  for seagrasses leaves. 
After calculating the uptake, the concentrations of ammonia, 
nitrate and phosphate in the water column are updated. This module is not calculating the sources-sinks terms
for roots.


Updates done to ModuleBenthicEcology:
Sources and sinks terms for seagrasses were included in a subroutine named as ComputeSeagrasses.


Several modifications were done to the subroutines for benthic feeders (filter feeders and deposit feeders) 
to better represent the dependence of ingestion rates on the cohesive sediment concentration and oxygen concentration.

Updates were done to ModuleInterface and ModuleInterfacesedimentWater, in order to transfer information between modules. In moduleInterface a new subroutine was added for the unfolding of R8 3D arrays.

ModuleSedimentProperties was modified to include the calculation of seagrasses roots distribution over the sediment column

ModuleWaterProperties was modified to include the calculation of seagrasses leaves distribution over the water column

ModuleGlobalData was edited to include all the new properties 

git-svn-id: https://mohid.svn.codeplex.com/svn@81362 81fbe566-5dc4-48c1-bdea-7421811ca204